Refactor database schema organization by splitting schema definitions into separate files for accounts, categories, and expenses
