+++ src/lib/server/db/schema/accounts.ts
@@ -1,0 +1,34 @@
+import { createId } from '@paralleldrive/cuid2';
+import { relations } from 'drizzle-orm';
+import { real, sqliteTable, text } from 'drizzle-orm/sqlite-core';
+import { createInsertSchema, createSelectSchema, createUpdateSchema } from 'drizzle-typebox';
+import { expenses } from './expenses';
+
+// ACCOUNTS
+export const accounts = sqliteTable('accounts', {
+	id: text()
+		.$defaultFn(() => createId())
+		.notNull()
+		.primaryKey(),
+	name: text('name').notNull(),
+	balance: real('balance').notNull(),
+	createdAt: text('created_at')
+		.notNull()
+		.$defaultFn(() => new Date().toISOString())
+		.$onUpdateFn(() => new Date().toISOString()),
+	updatedAt: text('updated_at')
+		.notNull()
+		.default(new Date().toISOString())
+		.$onUpdateFn(() => new Date().toISOString())
+});
+
+export const accountInsertSchema = createInsertSchema(accounts);
+export const accountSelectSchema = createSelectSchema(accounts);
+export const accountUpdateSchema = createUpdateSchema(accounts);
+
+export type InsertAccount = typeof accounts.$inferInsert;
+export type SelectAccount = typeof accounts.$inferSelect;
+
+export const accountsRelations = relations(accounts, ({ many }) => ({
+	expenses: many(expenses)
+}));
