+++ src/lib/server/db/schema/categories.ts
@@ -1,0 +1,21 @@
+import { createId } from '@paralleldrive/cuid2';
+import { sqliteTable, text } from 'drizzle-orm/sqlite-core';
+
+export const categories = sqliteTable('categories', {
+	id: text()
+		.$defaultFn(() => createId())
+		.notNull()
+		.primaryKey(),
+	name: text('name').notNull(),
+	createdAt: text('created_at')
+		.notNull()
+		.default(new Date().toISOString())
+		.$onUpdateFn(() => new Date().toISOString()),
+	updatedAt: text('updated_at')
+		.notNull()
+		.default(new Date().toISOString())
+		.$onUpdateFn(() => new Date().toISOString())
+});
+
+export type InsertCategory = typeof categories.$inferInsert;
+export type SelectCategory = typeof categories.$inferSelect;
