+++ src/lib/server/db/schema/index.ts
@@ -1,0 +1,3 @@
+export { accounts } from './accounts';
+export { categories } from './categories';
+export { expenses } from './expenses';
